fix timeout model

+++ database/src/models.rs
@@ -23,13 +23,13 @@ pub struct List {
     pub created_at: DateTime<Utc>,
     pub name: String,
     pub description: Option<String>,
-    pub timeout_seconds: Option<i64>,
+    pub timeout_seconds: Option<i32>,
 }
 
 pub struct ListUpdate {
     pub name: String,
     pub description: Option<String>,
-    pub timeout_seconds: Option<i64>,
+    pub timeout_seconds: Option<i32>,
 }
 
 #[derive(FromRow, Clone)]
